Add a CMake based compilation option
Adds a CMakeList.txt to make build the project via CMake. Also includes
a dummy hello world program to test on. As in CMake the install command
is dedicated to installing the produced software, renamed the install
command from the Makefile to install_deps

This also updates the README to provide guidance on how to compile with
 Cmake.